fix(kiloclaw): make kilo-exa compatible with openclaw 2026.4.5
Replace text-runtime helper imports with local normalization utilities so web_search no longer fails on runtimes that do not export normalizeOptionalLowercaseString. Align tests by removing the text-runtime stub alias and adding freshness normalization coverage.
