--- services/kiloclaw/plugins/kiloclaw-customizer/test/stubs/text-runtime.ts
@@ -1,12 +1,0 @@
-export function normalizeOptionalString(value: unknown): string | undefined {
-  if (typeof value !== 'string') {
-    return undefined;
-  }
-  const normalized = value.trim();
-  return normalized ? normalized : undefined;
-}
-
-export function normalizeOptionalLowercaseString(value: unknown): string | undefined {
-  const normalized = normalizeOptionalString(value);
-  return normalized ? normalized.toLowerCase() : undefined;
-}
